// sales-api/src/modules/queue-client/interfaces/queue-names.ts
export enum QueueName {
  EMAIL = 'email-queue',
  PDF_GENERATION = 'pdf-generation-queue',
  XML_PROCESSING = 'xml-processing-queue',
  CFDI_TIMBRADO = 'cfdi-timbrado-queue',
  NOTIFICATION = 'notification-queue',
  REPORT_GENERATION = 'report-generation-queue',
  INVENTORY_UPDATE = 'inventory-update-queue',
  ACCOUNTING = 'accounting-queue'
}